+++ src/inputValidation.js
@@ -155,6 +155,22 @@ const validateAgencyInputs = (agency, instance, agencyData) => {
 };
 
 /*
+      Warns user if National Aquifer Code is not 10 characters long. 
+      Prevents users from exceeding 1000 aquifer codes.
+    */
+const validateNatAquiferInputs = (natAquifer, instance) => {
+  let regex = /^(((\w){10}),)*((\w){10})$/;
+
+  if (instance.$store.getters.natAquiferActive) {
+    if (!natAquifer.replace(/\s/g, "").match(regex)) {
+      return "National Aquifer Code format is invalid. You may specify up to 1000 National Aquifer codes, separated by commas.";
+    }
+  }
+
+  return true;
+};
+
+/*
       Ensures that if both min and max values are present, the min value is less than the max value.
       Checks both hole and well depths.
     */
@@ -282,6 +298,24 @@ const validateFormInputs = instance => {
   );
   if (!(agencyStatus === true)) {
     notify(agencyStatus);
+    return false;
+  }
+
+  let natAquiferSiteStatus = validateNatAquiferInputs(
+    instance.$store.getters.natAquifer,
+    instance
+  );
+  if (!(natAquiferSiteStatus === true)) {
+    notify(natAquiferSiteStatus);
+    return false;
+  }
+
+  let locAquiferSiteStatus = validateLocAquiferInputs(
+    instance.$store.getters.locAquifer,
+    instance
+  );
+  if (!(locAquiferSiteStatus === true)) {
+    notify(locAquiferSiteStatus);
     return false;
   }